refactor(csharp): move ComplexTypeParser to its own ArrowTypeParser.cs
Promote the parser to a dedicated file with a clearer name that
reflects what it does (parses SQL type strings to Arrow types) without
overloading the "ComplexType" prefix shared with the serializer.

  csharp/src/ArrowTypeParser.cs                    (new)
  csharp/src/ComplexTypeSerializingStream.cs       (-184 lines)
  csharp/src/StatementExecution/StatementExecutionStatement.cs  (calls updated)
  csharp/test/Unit/StatementExecution/ParseComplexTypeTests.cs  (refs updated)

No logic change. 15 unit + 8 E2E permutations still pass.
